+++ NEWS.md
@@ -7,10 +7,14 @@ None
 1. Fix #397 - Add function `ts_geometric_brownian_motion()`
 2. Fix #402 - Add function `ts_brownian_motion_augment()`
 3. Fix #403 - Add function `ts_geometric_brownian_motion_augment()`
+4. Fix #404 - Add function `ts_brownian_motion_plot()`
 
 ## Minor Fixes and Improvements
 1. Fix #395 - Update and optimize `ts_brownian_motion()` 49x speedup by way of
 vectorization.
+2. Fix #412 - Update all brownian motion functions to have an attribute of 
+`.motion_type`
+3. Fix #411 - Drop the invisible return for `ts_vva_plot()`
 
 # healthyR.ts 0.2.6